+++ src/lib.rs
@@ -66,6 +66,7 @@ mod network_manager;
 mod packets;
 mod properties;
 pub mod reply_options;
+mod reason_codes;
 mod session_state;
 pub mod types;
 mod varint;
@@ -73,6 +74,7 @@ mod will;
 
 pub use properties::Property;
 pub use reply_options::ReplyOptions;
+pub use reason_codes::ReasonCode;
 
 pub use embedded_nal;
 pub use embedded_time;
@@ -122,21 +124,14 @@ pub enum Retain {
 /// Errors that are specific to the MQTT protocol implementation.
 #[derive(Debug, Copy, Clone, PartialEq)]
 pub enum ProtocolError {
-    Bounds,
-    DataSize,
-    Invalid,
-    Failed,
     PacketSize,
     MalformedPacket,
-    MalformedInteger,
-    UnknownProperty,
-    UnsupportedPacket,
     BufferSize,
     InvalidProperty,
     BadIdentifier,
     Unacknowledged,
     WrongQos,
-    Rejected(u8),
+    UnsupportedPacket,
     Serialization(crate::ser::Error),
     Deserialization(crate::de::Error),
 }
@@ -165,6 +160,12 @@ impl<E> From<crate::de::Error> for Error<E> {
     }
 }
 
+impl<E> From<ReasonCode> for Error<E> {
+    fn from(code: ReasonCode) -> Self {
+        Error::Failed(code)
+    }
+}
+
 /// Possible errors encountered during an MQTT connection.
 #[derive(Debug, PartialEq)]
 pub enum Error<E> {
@@ -174,7 +175,7 @@ pub enum Error<E> {
     Unsupported,
     ProvidedClientIdTooLong,
     NoResponseTopic,
-    Failed(u8),
+    Failed(ReasonCode),
     Protocol(ProtocolError),
     SessionReset,
     Clock(embedded_time::clock::Error),